// heft-plugins/heft-typescript-plugin-beta/src/types/typescript.d.ts
import type TTypescript from "typescript";

export { TTypescript };

export type ExtendedTypeScript = typeof TTypescript & {
  performance: {
    mark(name: string): void;
    measure(name: string, startMark?: string, endMark?: string): void;
    disable(): void;
    enable(): void;
    getDuration(name: string): number;
    getCount(name: string): number;
  };
  transpileOptionValueCompilerOptions: TTypescript.CommandLineOption[];
  getNewLineCharacter(
    compilerOptions: TTypescript.CompilerOptions | TTypescript.PrinterOptions
  ): string;
};